requirements.txt
onnxruntime
numpy==1.23.5
ruamel.yaml
colorama